Move DualView into its own header file
